pyproject.toml
[tool.poetry]
name = "api-taxi"
version = "0.1.0"
description = ""
authors = ["AlexM <alexmelnikoff@yandex.ru>"]
readme = "README.md"
packages = [{include = "api_taxi"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
pillow = "^9.5.0"
django-ckeditor = "^6.5.1"
gunicorn = "^20.1.0"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
